test(useRefEffect): consolidate and simplify regression tests
