+++ packages/react-simplikit/src/hooks/useRefEffect/useRefEffect.spec.tsx
@@ -1,10 +1,87 @@
 import { Fragment, startTransition, StrictMode, Suspense, useLayoutEffect } from 'react';
 import { act, render } from '@testing-library/react';
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
+
+import { renderHookSSR } from '../../_internal/test-utils/renderHookSSR.tsx';
 
 import { useRefEffect } from './useRefEffect.ts';
 
-describe('useRefEffect DOM lifecycle', () => {
+describe('useRefEffect', () => {
+  it('is safe on server side rendering', () => {
+    const callback = vi.fn();
+    renderHookSSR.serverOnly(() => useRefEffect(callback, []));
+
+    expect(callback).not.toHaveBeenCalled();
+  });
+
+  it('should call the callback when a new element is set', async () => {
+    const mockCallback = vi.fn();
+    const { result } = await renderHookSSR(() => useRefEffect(mockCallback, []));
+
+    const mockElement = document.createElement('div');
+    result.current(mockElement);
+
+    expect(mockCallback).toHaveBeenCalledTimes(1);
+    expect(mockCallback).toHaveBeenCalledWith(mockElement);
+  });
+
+  it('should call the cleanup function when the element changes', async () => {
+    const mockCleanup = vi.fn();
+    const mockCallback = vi.fn(() => mockCleanup);
+    const { result } = await renderHookSSR(() => useRefEffect(mockCallback, []));
+
+    const mockElement1 = document.createElement('div');
+    result.current(mockElement1);
+
+    const mockElement2 = document.createElement('div');
+    result.current(mockElement2);
+
+    expect(mockCallback).toHaveBeenCalledTimes(2);
+    expect(mockCleanup).toHaveBeenCalledTimes(1);
+  });
+
+  it('should not call the callback when setting null', async () => {
+    const mockCallback = vi.fn();
+    const { result } = await renderHookSSR(() => useRefEffect(mockCallback, []));
+
+    result.current(null);
+
+    expect(mockCallback).not.toHaveBeenCalled();
+  });
+
+  it('should call the cleanup function when setting null', async () => {
+    const mockCleanup = vi.fn();
+    const mockCallback = vi.fn(() => mockCleanup);
+    const { result } = await renderHookSSR(() => useRefEffect(mockCallback, []));
+
+    const mockElement = document.createElement('div');
+    result.current(mockElement);
+
+    result.current(null);
+
+    expect(mockCleanup).toHaveBeenCalledTimes(1);
+  });
+
+  it('should respect dependencies and re-initialize when they change', async () => {
+    const mockCallback1 = vi.fn();
+    const mockCallback2 = vi.fn();
+    let callback = mockCallback1;
+
+    const { result, rerender } = await renderHookSSR(() => useRefEffect(callback, [callback]));
+
+    const mockElement = document.createElement('div');
+    result.current(mockElement);
+
+    expect(mockCallback1).toHaveBeenCalledTimes(1);
+
+    callback = mockCallback2;
+    rerender();
+
+    result.current(mockElement);
+
+    expect(mockCallback2).toHaveBeenCalledTimes(1);
+  });
+
   it.each([false, true])('uses the new dependency value during ref attachment (StrictMode: %s)', strict => {
     const events: string[] = [];
 
@@ -71,49 +148,27 @@ describe('useRefEffect DOM lifecycle', () => {
     expect(events.at(-1)).toBe('cleanup:B:second');
   });
 
-  it('attaches with the new dependencies when a previously absent element appears', () => {
-    const events: string[] = [];
-
-    function Component({ value, visible }: { value: string; visible: boolean }) {
-      const ref = useRefEffect<HTMLDivElement>(() => {
-        events.push(`setup:${value}`);
-        return () => events.push(`cleanup:${value}`);
-      }, [value]);
-
-      return visible ? <div ref={ref} /> : null;
-    }
-
-    const { rerender, unmount } = render(<Component value="A" visible={false} />);
-    rerender(<Component value="B" visible />);
-    expect(events).toEqual(['setup:B']);
-
-    rerender(<Component value="C" visible={false} />);
-    expect(events).toEqual(['setup:B', 'cleanup:B']);
-    unmount();
-    expect(events).toHaveLength(2);
-  });
-
-  it.each([false, true])('does not expose callbacks from a suspended render (deps change: %s)', async changeDeps => {
+  it('ignores the callback from a render that did not commit', async () => {
     const events: string[] = [];
     const renderedValues: string[] = [];
     const pending = new Promise<never>(() => {});
-    let attach: (element: HTMLDivElement | null) => void = () => {};
+    let attachCommittedRef: (element: HTMLDivElement | null) => void = () => {};
 
     function Component({ value }: { value: string }) {
       const ref = useRefEffect<HTMLDivElement>(() => {
         events.push(`setup:${value}`);
         return () => events.push(`cleanup:${value}`);
-      }, [changeDeps ? value : 'stable']);
+      }, []);
 
       useLayoutEffect(
         function captureCommittedRef() {
-          attach = ref;
+          attachCommittedRef = ref;
         },
         [ref]
       );
 
       renderedValues.push(value);
-      if (value === 'B') {
+      if (value === 'uncommitted') {
         throw pending;
       }
 
@@ -122,7 +177,7 @@ describe('useRefEffect DOM lifecycle', () => {
 
     const { rerender, getByText } = render(
       <Suspense fallback="Loading">
-        <Component value="A" />
+        <Component value="committed" />
       </Suspense>
     );
 
@@ -130,17 +185,18 @@ describe('useRefEffect DOM lifecycle', () => {
       startTransition(() => {
         rerender(
           <Suspense fallback="Loading">
-            <Component value="B" />
+            <Component value="uncommitted" />
           </Suspense>
         );
       });
     });
 
-    expect(renderedValues).toContain('B');
-    expect(getByText('A')).toBeVisible();
+    expect(renderedValues).toContain('uncommitted');
+    expect(getByText('committed')).toBeVisible();
     events.length = 0;
 
-    act(() => attach(document.createElement('div')));
-    expect(events).toEqual(['cleanup:A', 'setup:A']);
+    // Reattaching the committed ref must ignore the callback from the suspended render.
+    act(() => attachCommittedRef(document.createElement('div')));
+    expect(events).toEqual(['cleanup:committed', 'setup:committed']);
   });
 });
